requirements.txt
# ===================================================================
# tpRigToolkit-core requirements file
# ===================================================================

Qt.py
pytest
pytest-cov
coveralls
sphinx
sphinx_rtd_theme
python-json-logger
sentry-sdk
pyyaml
metayaml
tpDcc-libs-python
tpDcc-core
tpDcc-libs-qt
tpDcc-dccs-maya
tpDcc-tools-nameit
tpRigToolkit-config